implement new features

+++ src/app/components/setting/setting.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { sharedService } from '../../providers/shared.service';
 import { remote } from 'electron';
-let { dialog } = remote;
+import { MatDialog } from '@angular/material/dialog';
+import { AddingComponent } from '../adding/adding.component';
+let { dialog, screen } = remote;
 
 import * as $ from 'jquery';
 import 'jquery-ui/ui/widgets/sortable.js';
@@ -12,9 +14,11 @@ import 'jquery-ui/ui/widgets/sortable.js';
   styleUrls: ['./setting.component.scss']
 })
 export class SettingComponent implements OnInit {
+  public screenHeight: any;
 
   constructor(
     public _sharedService: sharedService,
+    public mdDialog: MatDialog,
   ) {
     if( !this._sharedService.endpoints ) {
       this._sharedService.endpoints = [];
@@ -22,6 +26,9 @@ export class SettingComponent implements OnInit {
   }
 
   ngOnInit() {
+    const electronScreen = screen;
+    const size = electronScreen.getPrimaryDisplay().workAreaSize;
+    this.screenHeight = size.height;
   }
 
   ngAfterViewInit() {
@@ -42,7 +49,38 @@ export class SettingComponent implements OnInit {
         message: 'You can not remove currently selected endpoint!'
       });
     } else {
-      this._sharedService.removeOne(endpointId);
+      let options  = {
+        type: 'question',
+        buttons: ['Yes', 'No'],
+        message: 'Do you really want to remove?'
+       }
+       dialog.showMessageBox(remote.getCurrentWindow(), options)
+        .then(res => {
+          if(res.response == 0) {
+            this._sharedService.removeOne(endpointId);
+          } else {
+            return;
+          }
+        });
+    }
+  }
+
+  editOne(endpointId: any) {
+    if( endpointId == this._sharedService.curOpenedId ) {
+      dialog.showMessageBox(remote.getCurrentWindow(), {
+        type: 'warning',
+        buttons: ['OK'],
+        title: 'Warning!',
+        message: 'You can not edit currently selected endpoint!'
+      });
+    } else {
+      this._sharedService.addingDlgRef = this.mdDialog.open(AddingComponent, {
+        width: '900px',
+        height: (this.screenHeight - 72 - 15) + 'px',
+        disableClose: true,
+        data: {endpointId: endpointId}
+      });
+      this._sharedService.isHomePage = false;
     }
   }
 }
